fixed types

+++ src/main/java/operator/com/operator/models/dto/ShipmentsDto.java
@@ -46,9 +46,9 @@ public class ShipmentsDto {
 
     @NotBlank(message = "It's required")
     @Min(0)
-    private Integer discount;
+    private Double discount;
 
     @NotBlank(message = "It's required")
     @Min(0)
-    private Integer payment;
+    private Double payment;
 }
